Move OptionGroupsResponse to response/blocks package. Use  ViewResponseBase instead of BlockKitView. Remove BlockKitView.
