+++ slack-base/src/main/java/com/hubspot/slack/client/models/interaction/blockkit/BlockKitLoadOptionsRequestIF.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import com.hubspot.immutables.style.HubSpotStyle;
 import com.hubspot.slack.client.models.interaction.InteractiveLoadOptionsRequestType;
+import com.hubspot.slack.client.models.response.views.ViewResponseBase;
 import com.hubspot.slack.client.models.teams.SlackTeam;
 import org.immutables.value.Value;
 
@@ -12,7 +13,7 @@ import org.immutables.value.Value;
 @JsonNaming(value = PropertyNamingStrategy.SnakeCaseStrategy.class)
 public interface BlockKitLoadOptionsRequestIF {
     String getValue();
-    BlockKitView getView();
+    ViewResponseBase getView();
     String getActionId();
     SlackTeam getTeam();
     String getToken();
