--- slack-base/src/main/java/com/hubspot/slack/client/models/interaction/blockkit/BlockKitViewIF.java
@@ -1,13 +1,0 @@
-package com.hubspot.slack.client.models.interaction.blockkit;
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategy;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import com.hubspot.immutables.style.HubSpotStyle;
-import org.immutables.value.Value;
-
-@Value.Immutable
-@HubSpotStyle
-@JsonNaming(value = PropertyNamingStrategy.SnakeCaseStrategy.class)
-public interface BlockKitViewIF {
-    String getCallbackId();
-}
